refactor: peer4 and multiformats
lib folder was missing (probably due to .gitignore having lib/ in it) so
I added it and included a peer4 implementation in it. I did a bit of
refactoring and created a multiformats module as well and updated
didService to use it for it's multibase encoding.

I also added a prettierrc and some dev dependencies for types.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>